Add warning window

+++ src/components/SearchAppBar.js
@@ -7,12 +7,12 @@ import {
 } from 'react-bootstrap-icons';
 
 export default function SearchAppBar() {
-  let { currentProjectIndex, projects } = useAppState();
+  let { currentProjectIndex } = useAppState();
   const dispatch = useStateDispatch();
 
   let [searchText, setSearchText] = React.useState("")
   let [showQuery, setShowQuery] = React.useState(false)
-  let [queryList, setQueryList] = React.useState("")
+  let [queryList, setQueryList] = React.useState([])
 
 
   const delay = async (ms) => {
@@ -22,34 +22,35 @@ export default function SearchAppBar() {
 
   function SearchList() {
 
-    // let currentProject = projects[currentProjectIndex]
+    console.log({ queryList })
+    if (queryList.length > 0) {
+      let inactiveProjects = queryList.filter((item) => item.id !== "data_" + currentProjectIndex);
+      let queryListSearch = inactiveProjects.filter((item) => item.data.name.includes(searchText));
 
-    // console.log(Object.keys(localStorage))
-
-    // let inactiveProjects = queryList.filter((item) => item.id !== "data_" + currentProject);
-    // let queryList = inactiveProjects.filter((item) => item.includes(searchText));
-
-    return (<>
-      {showQuery === true ?
-        <div className='dropdown-content'>
-          {queryList.map(item => (
-            <div key={item.id} className='project-search-item'>
-              <SearchItem key={item.id} project={item} />
-              <Trash
-                onClick={() => {
-                  dispatch({
-                    type: 'delete-project',
-                    project: item
-                  });
-                }}
-                size={12} />
-            </div>
-          ))}
-        </div>
-        :
-        null}
-    </>
-    );
+      return (<>
+        {showQuery === true ?
+          <div className='dropdown-content'>
+            {queryListSearch.map(item => (
+              <div key={item.id} className='project-search-item'>
+                <SearchItem key={item.id} project={item} />
+                <Trash
+                  onClick={() => {
+                    dispatch({
+                      type: 'delete-project',
+                      project: item
+                    });
+                  }}
+                  size={12} />
+              </div>
+            ))}
+          </div>
+          :
+          null}
+      </>
+      );
+    } else {
+      return (<></>)
+    }
   }
 
   function SearchItem({ project }) {
@@ -77,11 +78,11 @@ export default function SearchAppBar() {
     let projectKeys = Object.keys(localStorage).filter((project) => project.includes('data_'))
 
     let projectData = projectKeys.map((key) => (
-                                    {
-                                      id: key,
-                                      data: JSON.parse(localStorage.getItem(key))
-                                    }
-                                    ))
+      {
+        id: key,
+        data: JSON.parse(localStorage.getItem(key))
+      }
+    ))
 
     // let projectNames = projectData.map((data) => data.name)
     // console.log(projectData)
